requirements.txt
requests
beautifulsoup4
pyyaml
python-dateutil
python-dotenv
